requirements.txt
python-chess>=1.999.0
numpy>=1.24.3
pygame>=2.5.2
pytest>=7.4.3
black>=23.11.0
pylint>=3.0.2
tqdm>=4.66.1
pandas>=2.1.3
matplotlib>=3.8.2
Pillow>=10.0.0
stockfish>=3.28.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.2